# Reuse IO::Sized instances in ResultSet