--- a/src/pg/result_set.cr
+++ b/src/pg/result_set.cr
@@ -15,6 +15,7 @@ class PG::ResultSet < ::DB::ResultSet
     @column_index = -1 # The current column
     @end = false       # Did we read all the rows?
     @rows_affected = 0_i64
+    @sized_io = IO::Sized.new(conn.soc, 1)
   end
 
   protected def conn
@@ -138,15 +139,15 @@ class PG::ResultSet < ::DB::ResultSet
   end
 
   private def safe_read(col_bytesize)
-    sized_io = IO::Sized.new(conn.soc, col_bytesize)
+    @sized_io.read_remaining = col_bytesize.to_u64
 
     begin
-      yield sized_io
+      yield @sized_io
     ensure
       # An exception might happen while decoding the value:
       # 1. Make sure to skip the column bytes
       # 2. Make sure to increment the column index
-      conn.soc.skip(sized_io.read_remaining) if sized_io.read_remaining > 0
+      conn.soc.skip(@sized_io.read_remaining) if @sized_io.read_remaining > 0
       @column_index += 1
     end
   end