@straight-shoota 2021-03-18
I'd suggest to make that property writable in the stdlib implementation. This is a perfectly valid use case and I would imagine this isn't that uncommon.