@will 2021-03-18
Great find, thank you for the investigation and PR! If you put together a PR for the crystal stdlib as @straight-shoota suggests please link it here.

Given how small the change will be, I wouldn't be opposed to back-porting it here in a monkey patch with a conditional on crystal version. Especially since it'd just be a temporary measure. But others may disagree since it is a little gross I suppose.