@jgaskins 2021-03-18
Well, then, brb making a PR to Crystal 😄