@jgaskins 2021-03-19
@will I've updated it to be a monkeypatch on `IO::Sized` and included a comment about why it's there, including a link to [the Crystal PR](https://github.com/crystal-lang/crystal/pull/10520). We have a backlink to the Crystal PR just above this comment, as well. I'll go ahead and mark this one as ready for review.